backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Supabase (includes authentication)
supabase==2.0.3
postgrest-py==0.10.6

# Gemini AI
google-generativeai==0.3.2

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2
requests==2.31.0

# Development
black==24.1.1
pytest==7.4.4
httpx==0.25.2
